Fix build types
