+++ .eslintrc.json
@@ -15,6 +15,8 @@
         }
     },
     "rules": {
+        "spaced-comment": [1, "always", { "markers": ["/"] }],
+        "@typescript-eslint/triple-slash-reference": [1, { "path": "always" }],
         "@typescript-eslint/consistent-type-imports": [1, { "disallowTypeAnnotations": false }],
         "@typescript-eslint/no-parameter-properties": 1,
         "@typescript-eslint/type-annotation-spacing": 1,
