+++ bundles/pixi.js-legacy/package.json
@@ -38,7 +38,8 @@
     "url": "https://github.com/pixijs/pixi.js.git"
   },
   "files": [
-    "dist/",
+    "lib",
+    "dist",
     "*.d.ts"
   ],
   "funding": {
