+++ bundles/pixi.js-node/package.json
@@ -26,7 +26,7 @@
     }
   },
   "files": [
-    "dist",
+    "lib",
     "*.d.ts"
   ],
   "engines": {
