+++ bundles/pixi.js-webworker/package.json
@@ -31,6 +31,7 @@
     }
   },
   "files": [
+    "lib",
     "dist",
     "*.d.ts"
   ],
