+++ package.json
@@ -23,11 +23,11 @@
     "lint": "eslint --ext .js --ext .ts test bundles packages scripts --ignore-path .gitignore --max-warnings 0",
     "lintfix": "run-s \"lint -- --fix\"",
     "types": "tsc -noEmit",
-    "build": "run-s clean:build build:packages",
-    "build:packages": "rollup -c",
+    "prebuild": "run-s clean:build",
+    "build": "rollup -c",
     "prebuild:types": "run-s clean:types clean:out",
-    "build:types": "tsc -p tsconfig.types.json",
-    "postbuild:types": "ts-node scripts/buildTypes.ts",
+    "build:types": "tsc -p tsconfig.types.json && ts-node scripts/fixTypes.ts",
+    "postbuild:types": "copyfiles -u 1 \"out/**/*\" . && run-s clean:out",
     "watch": "rollup -cw",
     "dist": "run-s lint types build:packages build:types docs",
     "dist-validate": "workspaces-run --only=\"@pixi/node\" -- npm test",
@@ -48,8 +48,6 @@
     "npm": ">=7"
   },
   "devDependencies": {
-    "@microsoft/api-extractor": "~7.20.1",
-    "@pixi-build-tools/api-extractor-lerna-monorepo": "~1.1.0",
     "@pixi/eslint-config": "^4.0.1",
     "@pixi/webdoc-template": "^1.5.4",
     "@rollup/plugin-commonjs": "^15.1.0",
@@ -61,7 +59,7 @@
     "cross-env": "^5.2.0",
     "electron": "^12.0.0",
     "esbuild": "^0.14.48",
-    "eslint": "^7.2.0",
+    "eslint": "^8.25.0",
     "eslint-plugin-jsdoc": "^39.2.9",
     "glob": "^7.1.3",
     "http-server": "^14.1.1",
