+++ packages/app/src/index.ts
@@ -1,3 +1,4 @@
+/// <reference path="../global.d.ts" />
 export * from './Application';
 export * from './ResizePlugin';
