+++ packages/extensions/package.json
@@ -29,7 +29,7 @@
     "access": "public"
   },
   "files": [
-    "dist",
+    "lib",
     "*.d.ts"
   ]
 }
